Host IFC functionality
Add a component that shows lists of network and storage interfaces on the host.
Clicking on the interfaces should control the host interfaces in the workspace

Dragging the host interfaces out of the workspace will delete them

Host interface from project database does not have any mapping. A cookie is created
to remember mapping for eg. eth0 -> host-00123